Add tomorrow theme
To make the Vim Powerline work with the correct colors I had to change my Terminal.app color settings to xterm-256. I liked ansi and I hope to find a way to go back to using ansi but for now I have to get used to xterm-256.

I also forgot to make the powerline to always appear in my last commit. Now it's always visible.

Signed-off-by: Patrick Kontschak <patrick.kontschak@gmail.com>