Add key bindings for split view
w is now creating a new vertical split view and puts the focus on the new view.

I also disabled the arrow keys in both normal and insert mode to force myself to always use hjkl to navigate and use the insert mode only when I have to edit text, not for navigation!

Signed-off-by: Patrick Kontschak <patrick.kontschak@gmail.com>