Modify cmd+ctrl+left binding
I wasn't very happy with the cmd+ctrl+left binding. There was a big gap inbetween 2 windows if each of them had cmd+ctrl+left or cmd+ctrl+right applied to them.

I set it so that right would be a small strip and left would fill the rest of the screen.

Signed-off-by: Patrick Kontschak <patrick.kontschak@gmail.com>